Fixing equal name bug
The Program could not be compiled when an example with the same name is
already open and one of the programs with the same name is closed. This
causes the ProgramController to remove its link to the program and stage
and the program will eventually crash. If now an example is opened with
a name that already exists, the existing instance will be closed. More
options like saving it with another name will be implemented later as
the example will overwrite the existing file.